Decouple configuration object from XML loader
