+++ src/Util/PHP/Template/TestCaseClass.tpl
@@ -1,6 +1,6 @@
 <?php
 use SebastianBergmann\CodeCoverage\CodeCoverage;
-use PHPUnit\TextUI\Configuration\Configuration;
+use PHPUnit\TextUI\Configuration\Registry;
 use PHPUnit\TextUI\Configuration\PhpHandler;
 
 if (!defined('STDOUT')) {
@@ -84,7 +84,7 @@ function __phpunit_run_isolated_test()
 $configurationFilePath = '{configurationFilePath}';
 
 if ('' !== $configurationFilePath) {
-    $configuration = Configuration::getInstance($configurationFilePath);
+    $configuration = Registry::getInstance()->get($configurationFilePath);
 
     (new PhpHandler)->handle($configuration->php());
